refactor and test fft
