+++ halo2_proofs/src/arithmetic.rs
@@ -9,7 +9,9 @@ use group::{
 };
 pub use pasta_curves::arithmetic::*;
 
+mod fft;
 mod multiexp;
+use fft::{parallel_fft, serial_fft, swap_bit_reverse};
 use multiexp::{multiexp_serial, small_multiexp};
 
 /// Performs a multi-exponentiation operation.
@@ -78,38 +80,6 @@ pub fn best_ifft<G: Group>(a: &mut [G], omega_inv: G::Scalar, log_n: u32, diviso
     });
 }
 
-/// This performs serial butterfly arithmetic
-pub(crate) fn serial_fft<G: Group>(a: &mut [G], n: usize, log_n: u32, twiddles: &[G::Scalar]) {
-    let mut chunk = 2_usize;
-    let mut twiddle_chunk = (n / 2) as usize;
-    for _ in 0..log_n {
-        a.chunks_mut(chunk).for_each(|coeffs| {
-            let (left, right) = coeffs.split_at_mut(chunk / 2);
-            butterfly_arithmetic(left, right, twiddle_chunk, twiddles)
-        });
-        chunk *= 2;
-        twiddle_chunk /= 2;
-    }
-}
-
-/// This perform recursive butterfly arithmetic
-pub fn parallel_fft<G: Group>(a: &mut [G], n: usize, twiddle_chunk: usize, twiddles: &[G::Scalar]) {
-    if n == 2 {
-        let t = a[1];
-        a[1] = a[0];
-        a[0].group_add(&t);
-        a[1].group_sub(&t);
-    } else {
-        let (left, right) = a.split_at_mut(n / 2);
-        rayon::join(
-            || parallel_fft(left, n / 2, twiddle_chunk * 2, twiddles),
-            || parallel_fft(right, n / 2, twiddle_chunk * 2, twiddles),
-        );
-
-        butterfly_arithmetic(left, right, twiddle_chunk, twiddles)
-    }
-}
-
 /// This evaluates a provided polynomial (in coefficient form) at `point`.
 pub fn eval_polynomial<F: Field>(poly: &[F], point: F) -> F {
     // TODO: parallelize?
@@ -154,44 +124,6 @@ where
     }
 
     q
-}
-
-/// This performs bit reverse permutation over `[G]`
-fn swap_bit_reverse<G: Group>(a: &mut [G], n: usize, log_n: u32) {
-    assert!(log_n <= 64);
-    let diff = 64 - log_n;
-    for i in 0..n as u64 {
-        let ri = i.reverse_bits() >> diff;
-        if i < ri {
-            a.swap(ri as usize, i as usize);
-        }
-    }
-}
-
-/// This performs butterfly arithmetic with two `G` array
-fn butterfly_arithmetic<G: Group>(
-    left: &mut [G],
-    right: &mut [G],
-    twiddle_chunk: usize,
-    twiddles: &[G::Scalar],
-) {
-    // case when twiddle factor is one
-    let t = right[0];
-    right[0] = left[0];
-    left[0].group_add(&t);
-    right[0].group_sub(&t);
-
-    left.iter_mut()
-        .zip(right.iter_mut())
-        .enumerate()
-        .skip(1)
-        .for_each(|(i, (a, b))| {
-            let mut t = *b;
-            t.group_scale(&twiddles[i * twiddle_chunk]);
-            *b = *a;
-            a.group_add(&t);
-            b.group_sub(&t);
-        });
 }
 
 /// Returns coefficients of an n - 1 degree polynomial given a set of n points
